requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pillow>=10.0.0
scipy>=1.10.0
scikit-image>=0.21.0
tqdm>=4.65.0
tensorboard>=2.13.0
lpips>=0.1.4
einops>=0.7.0
pyyaml>=6.0
matplotlib>=3.7.0
opencv-python>=4.8.0
